refactor!: make keys an optional input
Make keys an optional input that is set to None by default. If not specified, then all columns will be included. This is a breaking change, as previously users could pass in None for the keys input and get an empty dict in return. This behavior can be recovered by making keys an empty list instead of None.